Remove Ido from Customize; introduce Vertico
This commit also contains some tweaks added by Customize (add
'sudo-edit' and others to safe-packages).

We also had forgotten to remove the ':tangle no' setting from this
source block (something we probably did when the block contained an
Ido configuration.) So we've extended the commit with the removal of
this source-block setting.